[REF] charts: moved the chart components from pages/Dashboard to components/charts
