+++ fitness_app/src/components/charts/UserAnotherChart.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from "react";
 import { Chart } from "chart.js";
-import styles from "../styles/UserAnotherChart.module.css";
+import styles from "./UserAnotherChart.module.css";
 
 const UserAnotherChart = () => {
   const chartRef = useRef(null);
