+++ fitness_app/src/components/charts/UserCaloriesChart.module.css
