+++ fitness_app/src/components/charts/UserPerformanceChart.module.css
